Create ext directory for outside funcs, add uuid to ext dir
