+++ core/Cargo.toml
@@ -22,6 +22,7 @@ json = [
     "dep:pest",
     "dep:pest_derive",
 ]
+uuid = ["dep:uuid"]
 
 [target.'cfg(target_os = "linux")'.dependencies]
 io-uring = "0.6.1"
