+++ core/vdbe/mod.rs
@@ -24,7 +24,9 @@ pub mod sorter;
 mod datetime;
 
 use crate::error::{LimboError, SQLITE_CONSTRAINT_PRIMARYKEY};
-use crate::function::{AggFunc, FuncCtx, MathFunc, MathFuncArity, ScalarFunc};
+#[cfg(feature = "uuid")]
+use crate::ext::{exec_ts_from_uuid7, exec_uuid, exec_uuidblob, exec_uuidstr, ExtFunc, UuidFunc};
+use crate::function::{AggFunc, Func, FuncCtx, MathFunc, MathFuncArity, ScalarFunc};
 use crate::pseudo::PseudoCursor;
 use crate::schema::Table;
 use crate::storage::sqlite3_ondisk::DatabaseHeader;
@@ -46,10 +48,10 @@ use regex::Regex;
 use std::borrow::{Borrow, BorrowMut};
 use std::cell::RefCell;
 use std::collections::{BTreeMap, HashMap};
-use std::fmt::Display;
 use std::rc::{Rc, Weak};
 
 use uuid::{ContextV7, Timestamp, Uuid};
+
 pub type BranchOffset = i64;
 use macros::Description;
 pub type CursorID = usize;
@@ -2397,10 +2399,7 @@ impl Program {
                             | ScalarFunc::RandomBlob
                             | ScalarFunc::Sign
                             | ScalarFunc::Soundex
-                            | ScalarFunc::ZeroBlob
-                            | ScalarFunc::UuidStr
-                            | ScalarFunc::UuidBlob
-                            | ScalarFunc::Uuid7TS => {
+                            | ScalarFunc::ZeroBlob => {
                                 let reg_value = state.registers[*start_reg].borrow_mut();
                                 let result = match scalar_func {
                                     ScalarFunc::Sign => exec_sign(reg_value),
@@ -2415,26 +2414,10 @@ impl Program {
                                     ScalarFunc::RandomBlob => Some(exec_randomblob(reg_value)),
                                     ScalarFunc::ZeroBlob => Some(exec_zeroblob(reg_value)),
                                     ScalarFunc::Soundex => Some(exec_soundex(reg_value)),
-                                    ScalarFunc::UuidStr => Some(exec_uuidstr(reg_value)?),
-                                    ScalarFunc::UuidBlob => Some(exec_uuidblob(reg_value)?),
-                                    ScalarFunc::Uuid7TS => Some(exec_ts_from_uuid7(reg_value)),
                                     _ => unreachable!(),
                                 };
                                 state.registers[*dest] = result.unwrap_or(OwnedValue::Null);
                             }
-                            ScalarFunc::Uuid7 => match arg_count {
-                                0 => {
-                                    state.registers[*dest] =
-                                        exec_uuid(scalar_func, None).unwrap_or(OwnedValue::Null);
-                                }
-                                1 => {
-                                    let reg_value = state.registers[*start_reg].borrow();
-                                    state.registers[*dest] =
-                                        exec_uuid(scalar_func, Some(reg_value))
-                                            .unwrap_or(OwnedValue::Null);
-                                }
-                                _ => unreachable!(),
-                            },
                             ScalarFunc::Hex => {
                                 let reg_value = state.registers[*start_reg].borrow_mut();
                                 let result = exec_hex(reg_value);
@@ -2448,9 +2431,6 @@ impl Program {
                             }
                             ScalarFunc::Random => {
                                 state.registers[*dest] = exec_random();
-                            }
-                            ScalarFunc::Uuid4 | ScalarFunc::Uuid4Str => {
-                                state.registers[*dest] = exec_uuid(scalar_func, None)?;
                             }
                             ScalarFunc::Trim => {
                                 let reg_value = state.registers[*start_reg].clone();
@@ -2554,6 +2534,38 @@ impl Program {
                                 let replacement = &state.registers[*start_reg + 2];
                                 state.registers[*dest] = exec_replace(source, pattern, replacement);
                             }
+                        },
+                        Func::Extention(extfn) => match extfn {
+                            #[cfg(feature = "uuid")]
+                            ExtFunc::Uuid(uuidfn) => match uuidfn {
+                                UuidFunc::Uuid4 | UuidFunc::Uuid4Str => {
+                                    state.registers[*dest] = exec_uuid(uuidfn, None)?
+                                }
+                                UuidFunc::Uuid7 => match arg_count {
+                                    0 => {
+                                        state.registers[*dest] =
+                                            exec_uuid(uuidfn, None).unwrap_or(OwnedValue::Null);
+                                    }
+                                    1 => {
+                                        let reg_value = state.registers[*start_reg].borrow();
+                                        state.registers[*dest] = exec_uuid(uuidfn, Some(reg_value))
+                                            .unwrap_or(OwnedValue::Null);
+                                    }
+                                    _ => unreachable!(),
+                                },
+                                _ => {
+                                    // remaining accept 1 arg
+                                    let reg_value = state.registers[*start_reg].borrow();
+                                    state.registers[*dest] = match uuidfn {
+                                        UuidFunc::Uuid7TS => Some(exec_ts_from_uuid7(reg_value)),
+                                        UuidFunc::UuidStr => exec_uuidstr(reg_value).ok(),
+                                        UuidFunc::UuidBlob => exec_uuidblob(reg_value).ok(),
+                                        _ => unreachable!(),
+                                    }
+                                    .unwrap_or(OwnedValue::Null);
+                                }
+                            },
+                            _ => unreachable!(), // when more extension types are added
                         },
                         crate::function::Func::Math(math_func) => match math_func.arity() {
                             MathFuncArity::Nullary => match math_func {
@@ -3115,93 +3127,6 @@ fn exec_random() -> OwnedValue {
     getrandom::getrandom(&mut buf).unwrap();
     let random_number = i64::from_ne_bytes(buf);
     OwnedValue::Integer(random_number)
-}
-
-fn exec_uuid(var: &ScalarFunc, sec: Option<&OwnedValue>) -> Result<OwnedValue> {
-    match var {
-        ScalarFunc::Uuid4 => Ok(OwnedValue::Blob(Rc::new(
-            Uuid::new_v4().into_bytes().to_vec(),
-        ))),
-        ScalarFunc::Uuid4Str => Ok(OwnedValue::Text(Rc::new(Uuid::new_v4().to_string()))),
-        ScalarFunc::Uuid7 => {
-            let uuid = match sec {
-                Some(OwnedValue::Integer(ref seconds)) => {
-                    let ctx = ContextV7::new();
-                    if *seconds < 0 {
-                        // not valid unix timestamp, error or null?
-                        return Ok(OwnedValue::Null);
-                    }
-                    Uuid::new_v7(Timestamp::from_unix(ctx, *seconds as u64, 0))
-                }
-                _ => Uuid::now_v7(),
-            };
-            Ok(OwnedValue::Blob(Rc::new(uuid.into_bytes().to_vec())))
-        }
-        _ => unreachable!(),
-    }
-}
-
-fn exec_uuidstr(reg: &OwnedValue) -> Result<OwnedValue> {
-    match reg {
-        OwnedValue::Blob(blob) => {
-            let uuid = Uuid::from_slice(blob).map_err(|e| LimboError::ParseError(e.to_string()))?;
-            Ok(OwnedValue::Text(Rc::new(uuid.to_string())))
-        }
-        OwnedValue::Text(val) => {
-            let uuid = Uuid::parse_str(val).map_err(|e| LimboError::ParseError(e.to_string()))?;
-            Ok(OwnedValue::Text(Rc::new(uuid.to_string())))
-        }
-        OwnedValue::Null => Ok(OwnedValue::Null),
-        _ => Err(LimboError::ParseError(
-            "Invalid argument type for UUID function".to_string(),
-        )),
-    }
-}
-
-fn exec_uuidblob(reg: &OwnedValue) -> Result<OwnedValue> {
-    match reg {
-        OwnedValue::Text(val) => {
-            let uuid = Uuid::parse_str(val).map_err(|e| LimboError::ParseError(e.to_string()))?;
-            Ok(OwnedValue::Blob(Rc::new(uuid.as_bytes().to_vec())))
-        }
-        OwnedValue::Blob(blob) => {
-            let uuid = Uuid::from_slice(blob).map_err(|e| LimboError::ParseError(e.to_string()))?;
-            Ok(OwnedValue::Blob(Rc::new(uuid.as_bytes().to_vec())))
-        }
-        OwnedValue::Null => Ok(OwnedValue::Null),
-        _ => Err(LimboError::ParseError(
-            "Invalid argument type for UUID function".to_string(),
-        )),
-    }
-}
-
-fn exec_ts_from_uuid7(reg: &OwnedValue) -> OwnedValue {
-    let uuid = match reg {
-        OwnedValue::Blob(blob) => {
-            Uuid::from_slice(blob).map_err(|e| LimboError::ParseError(e.to_string()))
-        }
-        OwnedValue::Text(val) => {
-            Uuid::parse_str(val).map_err(|e| LimboError::ParseError(e.to_string()))
-        }
-        _ => Err(LimboError::ParseError(
-            "Invalid argument type for UUID function".to_string(),
-        )),
-    };
-    match uuid {
-        Ok(uuid) => OwnedValue::Integer(uuid_to_unix(uuid.as_bytes()) as i64),
-        // display error? sqlean seems to set value to null
-        Err(_) => OwnedValue::Null,
-    }
-}
-
-#[inline(always)]
-fn uuid_to_unix(uuid: &[u8; 16]) -> u64 {
-    ((uuid[0] as u64) << 40)
-        | ((uuid[1] as u64) << 32)
-        | ((uuid[2] as u64) << 24)
-        | ((uuid[3] as u64) << 16)
-        | ((uuid[4] as u64) << 8)
-        | (uuid[5] as u64)
 }
 
 fn exec_randomblob(reg: &OwnedValue) -> OwnedValue {
@@ -4969,203 +4894,5 @@ mod tests {
             exec_replace(&input_str, &pattern_str, &replace_str),
             expected_str
         );
-    }
-
-    #[test]
-    fn test_exec_uuid_v4blob() {
-        use super::{exec_uuid, ScalarFunc};
-        use uuid::Uuid;
-        let func = ScalarFunc::Uuid4;
-        let owned_val = exec_uuid(&func, None);
-        match owned_val {
-            Ok(OwnedValue::Blob(blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(&blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 4);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v4str() {
-        use super::{exec_uuid, ScalarFunc};
-        use uuid::Uuid;
-        let func = ScalarFunc::Uuid4Str;
-        let owned_val = exec_uuid(&func, None);
-        match owned_val {
-            Ok(OwnedValue::Text(v4str)) => {
-                assert_eq!(v4str.len(), 36);
-                let uuid = Uuid::parse_str(&v4str);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 4);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v7_now() {
-        use super::{exec_uuid, ScalarFunc};
-        use uuid::Uuid;
-        let func = ScalarFunc::Uuid7;
-        let owned_val = exec_uuid(&func, None);
-        match owned_val {
-            Ok(OwnedValue::Blob(blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(&blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 7);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v7_with_input() {
-        use super::{exec_uuid, ScalarFunc};
-        use uuid::Uuid;
-        let func = ScalarFunc::Uuid7;
-        let owned_val = exec_uuid(&func, Some(&OwnedValue::Integer(946702800)));
-        match owned_val {
-            Ok(OwnedValue::Blob(blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(&blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 7);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v7_now_to_timestamp() {
-        use super::{exec_ts_from_uuid7, exec_uuid, ScalarFunc};
-        use uuid::Uuid;
-        let func = ScalarFunc::Uuid7;
-        let owned_val = exec_uuid(&func, None);
-        match owned_val {
-            Ok(OwnedValue::Blob(ref blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 7);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-        let result = exec_ts_from_uuid7(&owned_val.expect("uuid7"));
-        if let OwnedValue::Integer(ref ts) = result {
-            let unixnow = (std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap()
-                .as_secs()
-                * 1000) as i64;
-            assert!(*ts >= unixnow - 1000);
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v7_to_timestamp() {
-        use super::{exec_ts_from_uuid7, exec_uuid, ScalarFunc};
-        use uuid::Uuid;
-        let func = ScalarFunc::Uuid7;
-        let owned_val = exec_uuid(&func, Some(&OwnedValue::Integer(946702800)));
-        match owned_val {
-            Ok(OwnedValue::Blob(ref blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 7);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-        let result = exec_ts_from_uuid7(&owned_val.expect("uuid7"));
-        assert_eq!(result, OwnedValue::Integer(946702800 * 1000));
-        if let OwnedValue::Integer(ts) = result {
-            let time = chrono::DateTime::from_timestamp(ts / 1000, 0);
-            assert_eq!(
-                time.unwrap(),
-                "2000-01-01T05:00:00Z"
-                    .parse::<chrono::DateTime<chrono::Utc>>()
-                    .unwrap()
-            );
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v4_str_to_blob() {
-        use super::{exec_uuid, exec_uuidblob, ScalarFunc};
-        use uuid::Uuid;
-        let owned_val = exec_uuidblob(
-            &exec_uuid(&ScalarFunc::Uuid4Str, None).expect("uuid v4 string to generate"),
-        );
-        match owned_val {
-            Ok(OwnedValue::Blob(blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(&blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 4);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v7_str_to_blob() {
-        use super::{exec_uuid, exec_uuidblob, exec_uuidstr, ScalarFunc};
-        use uuid::Uuid;
-        // convert a v7 blob to a string then back to a blob
-        let owned_val = exec_uuidblob(
-            &exec_uuidstr(&exec_uuid(&ScalarFunc::Uuid7, None).expect("uuid v7 blob to generate"))
-                .expect("uuid v7 string to generate"),
-        );
-        match owned_val {
-            Ok(OwnedValue::Blob(blob)) => {
-                assert_eq!(blob.len(), 16);
-                let uuid = Uuid::from_slice(&blob);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 7);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v4_blob_to_str() {
-        use super::{exec_uuid, exec_uuidstr, ScalarFunc};
-        use uuid::Uuid;
-        // convert a v4 blob to a string
-        let owned_val =
-            exec_uuidstr(&exec_uuid(&ScalarFunc::Uuid4, None).expect("uuid v7 blob to generate"));
-        match owned_val {
-            Ok(OwnedValue::Text(v4str)) => {
-                assert_eq!(v4str.len(), 36);
-                let uuid = Uuid::parse_str(&v4str);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 4);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
-    }
-
-    #[test]
-    fn test_exec_uuid_v7_blob_to_str() {
-        use super::{exec_uuid, exec_uuidstr, ScalarFunc};
-        use uuid::Uuid;
-        // convert a v7 blob to a string
-        let owned_val = exec_uuidstr(
-            &exec_uuid(&ScalarFunc::Uuid7, Some(&OwnedValue::Integer(123456789)))
-                .expect("uuid v7 blob to generate"),
-        );
-        match owned_val {
-            Ok(OwnedValue::Text(v7str)) => {
-                assert_eq!(v7str.len(), 36);
-                let uuid = Uuid::parse_str(&v7str);
-                assert!(uuid.is_ok());
-                assert_eq!(uuid.unwrap().get_version_num(), 7);
-            }
-            _ => panic!("exec_uuid did not return a Blob variant"),
-        }
     }
 }
